# Store equality and subset comparisons with one pool in a field