--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -1,5 +1,6 @@
 using PkgBenchmark
 using CategoricalArrays
+using Printf
 
 @benchgroup "isequal(A, v::String)" begin
     function sumequals(A::AbstractArray, v::Any)
@@ -46,4 +47,21 @@ end
 
     levs = [levels(ca); [@sprintf("id2%010d", k) for k in 1:1000]]
     @bench "levels! with many additional levels" levels!(ca, levs)
+end
+
+@benchgroup "repeated assignment" begin
+    function mycopy!(dest, src)
+        @inbounds for i in eachindex(dest, src)
+            dest[i] = src[i]
+        end
+    end
+
+    a = categorical(rand([@sprintf("id%010d", k) for k in 1:1000], 10000))
+    b = CategoricalArray{String}(undef, 10000)
+    c = categorical(rand([@sprintf("id%010d", k) for k in 1:1000], 10000))
+    d = categorical(rand([@sprintf("id%010d", k) for k in 1001:2000], 10000))
+
+    @bench "empty dest" mycopy!(b2, a) setup = b2=copy(b)
+    @bench "same levels dest" mycopy!(c2, a) setup = c2=copy(c)
+    @bench "many levels dest" mycopy!(d2, a) setup = d2=copy(d)
 end