--- a/test/07_levels.jl
+++ b/test/07_levels.jl
@@ -1,7 +1,7 @@
 module TestLevels
 using Test
 using CategoricalArrays
-using CategoricalArrays: DefaultRefType, levels!
+using CategoricalArrays: DefaultRefType, levels!, hashlevels
 
 @testset "CategoricalPool{Int} updates levels and order correctly" begin
     pool = CategoricalPool([2, 1, 3])
@@ -11,6 +11,9 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels(pool) == [2, 1, 3]
     @test all([levels(CategoricalValue(i, pool)) for i in 1:3] .=== Ref(levels(pool)))
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
 
     for rep in 1:3
         push!(pool, 4)
@@ -19,6 +22,9 @@ using CategoricalArrays: DefaultRefType, levels!
         @test length(pool) === 4
         @test levels(pool) == [2, 1, 3, 4]
         @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4)
+        @test pool.hash == hashlevels(levels(pool))
+        @test pool.equalto == 0
+        @test pool.subsetof == 0
         @test get(pool, 4) === DefaultRefType(4)
         @test pool[4] === CategoricalValue(4, pool)
     end
@@ -30,6 +36,9 @@ using CategoricalArrays: DefaultRefType, levels!
         @test length(pool) === 5
         @test levels(pool) == [2, 1, 3, 4, 0]
         @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5)
+        @test pool.hash == hashlevels(levels(pool))
+        @test pool.equalto == 0
+        @test pool.subsetof == 0
         @test get(pool, 0) === DefaultRefType(5)
         @test pool[5] === CategoricalValue(5, pool)
     end
@@ -41,6 +50,9 @@ using CategoricalArrays: DefaultRefType, levels!
         @test length(pool) === 7
         @test levels(pool) == [2, 1, 3, 4, 0, 10, 11]
         @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7)
+        @test pool.hash == hashlevels(levels(pool))
+        @test pool.equalto == 0
+        @test pool.subsetof == 0
         @test get(pool, 10) === DefaultRefType(6)
         @test get(pool, 11) === DefaultRefType(7)
         @test pool[6] === CategoricalValue(6, pool)
@@ -54,6 +66,9 @@ using CategoricalArrays: DefaultRefType, levels!
         @test length(pool) === 9
         @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13]
         @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9)
+        @test pool.hash == hashlevels(levels(pool))
+        @test pool.equalto == 0
+        @test pool.subsetof == 0
         @test get(pool, 12) === DefaultRefType(8)
         @test get(pool, 13) === DefaultRefType(9)
         @test pool[8] === CategoricalValue(8, pool)
@@ -71,12 +86,18 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels!(pool, levs) === pool
     @test levels(pool) == levs
     @test levels(pool) !== levs
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
 
     @test isa(pool.levels, Vector{Int})
     @test length(pool) === 11
     @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13, 15, 14]
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9,
                                 15=>10, 14=>11)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
     @test get(pool, 15) === DefaultRefType(10)
     @test get(pool, 14) === DefaultRefType(11)
     @test pool[10] === CategoricalValue(10, pool)
@@ -94,6 +115,9 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13, 15, 14, 20]
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9,
                                 15=>10, 14=>11, 20=>12)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
     @test get(pool, 20) === DefaultRefType(12)
 
     # get! with CategoricalValue adding new levels in conflicting order
@@ -114,6 +138,9 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13, 15, 14, 20, 100, 99]
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9,
                                 15=>10, 14=>11, 20=>12, 100=>13, 99=>14)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
     @test get(pool, 100) === DefaultRefType(13)
     @test get(pool, 99) === DefaultRefType(14)
 
@@ -126,6 +153,9 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13, 15, 14, 20, 100, 99]
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9,
                                 15=>10, 14=>11, 20=>12, 100=>13, 99=>14)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
 
     # get! with CategoricalValue from same pool
     @test get!(pool, pool[1]) === DefaultRefType(1)
@@ -135,6 +165,9 @@ using CategoricalArrays: DefaultRefType, levels!
     @test levels(pool) == [2, 1, 3, 4, 0, 10, 11, 12, 13, 15, 14, 20, 100, 99]
     @test pool.invindex == Dict(2=>1, 1=>2, 3=>3, 4=>4, 0=>5, 10=>6, 11=>7, 12=>8, 13=>9,
                                 15=>10, 14=>11, 20=>12, 100=>13, 99=>14)
+    @test pool.hash == hashlevels(levels(pool))
+    @test pool.equalto == 0
+    @test pool.subsetof == 0
 
     # get! with CategoricalValue conversion error
     v = CategoricalValue(1, CategoricalPool(["a", "b"]))
@@ -173,6 +206,7 @@ end
 
 @testset "issubset" begin
     pool1 = CategoricalPool(["a", "b", "c"])
+    pool1b = CategoricalPool(["a", "b", "c"])
     pool2 = CategoricalPool(["c", "a", "b"])
     pool3 = CategoricalPool(["a", "b", "c", "d"])
     pool4 = CategoricalPool(["a", "b"])
@@ -180,11 +214,48 @@ end
     pool6 = CategoricalPool(String[])
 
     @test issubset(pool1, pool1)
+    @test pool1.equalto == 0
+    @test pool1.subsetof == 0
+
     @test issubset(pool2, pool1)
+    @test pool1.equalto == pool2.equalto == 0
+    @test pool1.subsetof == 0
+    @test pool2.subsetof == objectid(pool1)
+
+    @test issubset(pool1b, pool1)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool1b.equalto == objectid(pool1)
+    @test pool1.subsetof == 0
+    @test pool2.subsetof == objectid(pool1)
+
     @test !issubset(pool3, pool1)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool3.equalto == 0
+    @test pool1.subsetof == pool3.subsetof == 0
+
+    @test issubset(pool1, pool3)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool3.equalto == 0
+    @test pool1.subsetof == objectid(pool3)
+    @test pool3.subsetof == 0
+
     @test issubset(pool4, pool1)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool4.equalto == 0
+    @test pool1.subsetof == objectid(pool3)
+    @test pool4.subsetof == objectid(pool1)
+
     @test !issubset(pool5, pool1)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool5.equalto == 0
+    @test pool1.subsetof == objectid(pool3)
+    @test pool5.subsetof == 0
+
     @test issubset(pool6, pool1)
+    @test pool1.equalto == objectid(pool1b)
+    @test pool6.equalto == 0
+    @test pool1.subsetof == objectid(pool3)
+    @test pool6.subsetof == objectid(pool1)
 end
 
 end