--- a/test/08_equality.jl
+++ b/test/08_equality.jl
@@ -68,6 +68,8 @@ using CategoricalArrays
 
     @test hash(pool1) == hash(pool1b)
     @test hash(pool1) != hash(pool2) != hash(pool3) != hash(pool4)
+    @test hash(pool1, UInt(1234)) == hash(pool1b, UInt(1234)) != hash(pool1)
+    @test hash(pool1, UInt(1234)) != hash(pool2, UInt(1234))
 
     poolnan1 = CategoricalPool([1.0, NaN])
     poolnan2 = CategoricalPool([1.0, NaN])