When comparing `CategoricalValue`s from different pools, checking whether pools are equal or whether one is a superset of the other is too costly to be done repeatedly for all values in an array. To fix this, store the result of the comparison in new fields, so that subsequent checks are fast. This means that the check only needs to be done the first time two pools are compared,
or the first time after a comparison with another pool was made (which is rare). For equality checks, this relies on checking that none of the two pools was modified since they were marked as equal. For subset checks, this relies on the fact that pools can only be added new levels, never lose levels. This approach is simpler and more thread-safe than keeping a global table.

This makes it possible to support relatively efficient comparisons between `CategoricalValue`s from different pools using `==` and `<`. Also implement `==`/`isequal` and `hash` on `CategoricalPool`, with `==` being equivalent to `isequal` (like for `Dict`) for performance and ignoring the orderedness of the pool (like `CategoricalValue`).

master:
```julia
using BenchmarkTools
using CategoricalArrays
using Printf

function mycopy!(dest, src)
    @inbounds for i in eachindex(dest, src)
        dest[i] = src[i]
    end
end

a = categorical(rand([@sprintf("id%010d", k) for k in 1:1000], 10000))
b = CategoricalArray{String}(undef, 10000)
c = categorical(rand([@sprintf("id%010d", k) for k in 1:1000], 10000))
d = categorical(rand([@sprintf("id%010d", k) for k in 1001:2000], 10000))


julia> @btime mycopy!(b2, a) setup = b2=copy(b)
  635.083 ms (10009 allocations: 347.25 KiB)

julia> @btime mycopy!(c2, a) setup = c2=copy(c)
  784.074 ms (10000 allocations: 312.50 KiB)

julia> @btime mycopy!(d2, a) setup = d2=copy(d)
  866.868 ms (10047 allocations: 508.67 KiB)
```

PR:
```julia
julia> @btime mycopy!(b2, a) setup = b2=copy(b)

  773.434 μs (10009 allocations: 347.28 KiB)

julia> @btime mycopy!(c2, a) setup = c2=copy(c)
  827.869 μs (10000 allocations: 312.50 KiB)

julia> @btime mycopy!(d2, a) setup = d2=copy(d)
  1.199 ms (10047 allocations: 508.70 KiB)
```