@nalimilan 2021-04-03
It's even faster with pointers!
```julia
julia> @btime mycopy!(b2, a) setup = b2=copy(b)
  223.029 μs (10009 allocations: 347.28 KiB)

julia> @btime mycopy!(c2, a) setup = c2=copy(c)
  187.238 μs (10000 allocations: 312.50 KiB)

julia> @btime mycopy!(d2, a) setup = d2=copy(d)
  1.174 ms (10047 allocations: 508.70 KiB)
```