@bkamins 2021-04-03
Assuming you know all places where pool can be changed as discussed :).