@nalimilan 2021-04-04
A few more `@inline` annotations improve performance a bit:
```julia
julia> @btime mycopy!(b2, a) setup = b2=copy(b)
  172.229 μs (10009 allocations: 347.28 KiB)

julia> @btime mycopy!(c2, a) setup = c2=copy(c)
  122.986 μs (10000 allocations: 312.50 KiB)

julia> @btime mycopy!(d2, a) setup = d2=copy(d)
  1.128 ms (10047 allocations: 508.70 KiB)
```